small verification

+++ Models/Order.h
@@ -60,7 +60,7 @@ public:
     void setObservation(std::string observation);
     std::string getObservation();
 
-    long long getDiff();//time difference between start and end
+    int getDiff();//time difference between start and end
 
     //general function for writing an order's details
     void writeAll();
